5) Write a C sharp program to swap two numbers.
5) Write a C sharp program to swap two numbers .

Here we swap two numbers by using a temporary variable called temp. (we can also swap without using third variable like temp)
logic is pretty straight forward.

we take two variable with values lets say a as 3 and b as  4
temp will be 0 by default .

we push 3 to temp...that makes a empty.
now since a is empty we push value of b (4) to a.
now b will be empty , so we finally push temp which now has 3 to b..

thus new value of a will be 4    and     b will be 5.